+++ src/core/gpu_sw.h
@@ -37,7 +37,7 @@ protected:
   };
 
   void ReadVRAM(u32 x, u32 y, u32 width, u32 height, void* buffer) override;
-  void FillVRAM(u32 x, u32 y, u32 width, u32 height, u16 color) override;
+  void FillVRAM(u32 x, u32 y, u32 width, u32 height, u32 color) override;
   void UpdateVRAM(u32 x, u32 y, u32 width, u32 height, const void* data) override;
   void CopyVRAM(u32 src_x, u32 src_y, u32 dst_x, u32 dst_y, u32 width, u32 height) override;
